Separate to github_status
